+++ integration_tests/suite/test_relocates.py
@@ -79,10 +79,10 @@ class TestRelocates(RealAsteriskIntegrationTest):
         callee.setChannelVar(variable='XIVO_USERUUID', value=callee_uuid)
         bridge.addChannel(channel=callee.id)
 
-        calld = self.make_calld(VALID_TOKEN)
+        self.calld_client.set_token(VALID_TOKEN)
 
         def channels_have_been_created_in_calld(caller_id, callee_id):
-            calls = calld.calls.list_calls(application=STASIS_APP, application_instance=STASIS_APP_INSTANCE)
+            calls = self.calld_client.calls.list_calls(application=STASIS_APP, application_instance=STASIS_APP_INSTANCE)
             channel_ids = [call['call_id'] for call in calls['items']]
             return (caller_id in channel_ids and callee_id in channel_ids)
 
@@ -119,9 +119,13 @@ class TestRelocates(RealAsteriskIntegrationTest):
         self.confd.set_users(MockUser(uuid=user_uuid, line_ids=[line_id], mobile='mobile-autoanswer'))
         self.confd.set_lines(MockLine(id=line_id, name=SOME_LINE_NAME, protocol='local', context='local'))
         token = self.given_user_token(user_uuid)
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
-        call = calld.calls.make_call_from_user(extension='dial-autoanswer', from_mobile=True, variables={'CALLEE_XIVO_USERUUID': user_uuid})
+        call = self.calld_client.calls.make_call_from_user(
+            extension='dial-autoanswer',
+            from_mobile=True,
+            variables={'CALLEE_XIVO_USERUUID': user_uuid}
+        )
 
         def bridged_channel(caller):
             try:
@@ -149,10 +153,10 @@ class TestRelocates(RealAsteriskIntegrationTest):
         token = self.given_user_token(user_uuid)
         self.confd.set_users(MockUser(uuid=user_uuid, line_ids=[line_id]))
         self.confd.set_lines(MockLine(id=line_id, name='recipient@local', protocol='local', context=SOME_CONTEXT))
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
         destination = 'line'
         location = {'line_id': line_id}
-        relocate = calld.relocates.create_from_user(initiator_channel_id, destination, location)
+        relocate = self.calld_client.relocates.create_from_user(initiator_channel_id, destination, location)
 
         return relocate, user_uuid, destination, location
 
@@ -163,10 +167,10 @@ class TestRelocates(RealAsteriskIntegrationTest):
         token = self.given_user_token(user_uuid)
         self.confd.set_users(MockUser(uuid=user_uuid, line_ids=[line_id]))
         self.confd.set_lines(MockLine(id=line_id, name='recipient_autoanswer@local', protocol='local', context=SOME_CONTEXT))
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
         destination = 'line'
         location = {'line_id': line_id}
-        relocate = calld.relocates.create_from_user(initiator_channel_id, destination, location)
+        relocate = self.calld_client.relocates.create_from_user(initiator_channel_id, destination, location)
 
         def relocate_waiting_relocated():
             assert_that(relocated_channel_id, self.c.is_talking(), 'relocated channel not talking')
@@ -184,11 +188,11 @@ class TestRelocates(RealAsteriskIntegrationTest):
         token = self.given_user_token(user_uuid)
         self.confd.set_users(MockUser(uuid=user_uuid, line_ids=[line_id]))
         self.confd.set_lines(MockLine(id=line_id, name='recipient_autoanswer@local', protocol='local', context=SOME_CONTEXT))
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
         destination = 'line'
         location = {'line_id': line_id}
         completions = ['api']
-        relocate = calld.relocates.create_from_user(initiator_channel_id, destination, location, completions)
+        relocate = self.calld_client.relocates.create_from_user(initiator_channel_id, destination, location, completions)
 
         def all_talking():
             assert_that(relocate['relocated_call'], self.c.is_talking(), 'relocated channel not talking')
@@ -206,10 +210,10 @@ class TestRelocates(RealAsteriskIntegrationTest):
         token = self.given_user_token(user_uuid)
         self.confd.set_users(MockUser(uuid=user_uuid, line_ids=[line_id]))
         self.confd.set_lines(MockLine(id=line_id, name='recipient_autoanswer@local', protocol='local', context=SOME_CONTEXT))
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
         destination = 'line'
         location = {'line_id': line_id}
-        relocate = calld.relocates.create_from_user(initiator_channel_id, destination, location)
+        relocate = self.calld_client.relocates.create_from_user(initiator_channel_id, destination, location)
 
         until.assert_(
             self.assert_relocate_is_completed,
@@ -249,9 +253,9 @@ class TestListUserRelocate(TestRelocates):
     def test_given_no_relocates_when_list_then_list_empty(self):
         user_uuid = SOME_USER_UUID
         token = self.given_user_token(user_uuid)
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
-        result = calld.relocates.list_from_user()
+        result = self.calld_client.relocates.list_from_user()
 
         assert_that(result['items'], empty())
 
@@ -259,9 +263,9 @@ class TestListUserRelocate(TestRelocates):
         user_uuid = SOME_USER_UUID
         token = self.given_user_token(user_uuid)
         relocate, user_uuid, destination, location = self.given_ringing_user_relocate()
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
-        result = calld.relocates.list_from_user()
+        result = self.calld_client.relocates.list_from_user()
 
         assert_that(result['items'], contains({
             'uuid': relocate['uuid'],
@@ -276,10 +280,10 @@ class TestListUserRelocate(TestRelocates):
     def test_given_one_completed_relocate_when_list_then_relocate_not_found(self):
         user_uuid = SOME_USER_UUID
         token = self.given_user_token(user_uuid)
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
         relocate, user_uuid = self.given_completed_user_relocate()
 
-        relocates = calld.relocates.list_from_user()
+        relocates = self.calld_client.relocates.list_from_user()
 
         assert_that(relocates['items'], not_(contains(has_entry('uuid', relocate['uuid']))))
 
@@ -287,9 +291,9 @@ class TestListUserRelocate(TestRelocates):
         relocate1, user_uuid1, _, __ = self.given_ringing_user_relocate()
         relocate2, user_uuid2, _, __ = self.given_ringing_user_relocate()
         token = self.given_user_token(user_uuid2)
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
-        result = calld.relocates.list_from_user()
+        result = self.calld_client.relocates.list_from_user()
 
         assert_that(result['items'], contains(has_entries({
             'uuid': relocate2['uuid'],
@@ -300,9 +304,9 @@ class TestGetUserRelocate(TestRelocates):
 
     def test_given_no_relocate_when_get_then_error_404(self):
         token = self.given_user_token(SOME_USER_UUID)
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
-        assert_that(calling(calld.relocates.get_from_user).with_args(relocate_uuid='not-found'),
+        assert_that(calling(self.calld_client.relocates.get_from_user).with_args(relocate_uuid='not-found'),
                     raises(CalldError).matching(has_properties({
                         'status_code': 404,
                         'error_id': 'no-such-relocate',
@@ -311,9 +315,9 @@ class TestGetUserRelocate(TestRelocates):
     def test_given_other_relocate_when_get_then_404(self):
         relocate, user_uuid, _, __ = self.given_ringing_user_relocate()
         token = self.given_user_token(SOME_USER_UUID)
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
-        assert_that(calling(calld.relocates.get_from_user).with_args(relocate['uuid']),
+        assert_that(calling(self.calld_client.relocates.get_from_user).with_args(relocate['uuid']),
                     raises(CalldError).matching(has_properties({
                         'status_code': 404,
                         'error_id': 'no-such-relocate',
@@ -322,9 +326,9 @@ class TestGetUserRelocate(TestRelocates):
     def test_given_relocate_when_get_then_result(self):
         relocate, user_uuid, _, __ = self.given_ringing_user_relocate()
         token = self.given_user_token(user_uuid)
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
-        result = calld.relocates.get_from_user(relocate['uuid'])
+        result = self.calld_client.relocates.get_from_user(relocate['uuid'])
 
         assert_that(result, has_entries({
             'uuid': relocate['uuid'],
@@ -338,17 +342,17 @@ class TestCreateUserRelocate(TestRelocates):
         self.confd.reset()
 
     def test_given_wrong_token_when_relocate_then_401(self):
-        calld = self.make_calld(INVALID_ACL_TOKEN)
+        self.calld_client.set_token(INVALID_ACL_TOKEN)
 
-        assert_that(calling(calld.relocates.create_from_user).with_args(SOME_CALL_ID, 'destination'),
+        assert_that(calling(self.calld_client.relocates.create_from_user).with_args(SOME_CALL_ID, 'destination'),
                     raises(CalldError).matching(has_property('status_code', 401)))
 
     def test_given_invalid_request_when_relocate_then_400(self):
         user_uuid = SOME_USER_UUID
         token = self.given_user_token(user_uuid)
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
-        assert_that(calling(calld.relocates.create_from_user).with_args(SOME_CALL_ID, 'wrong-destination'),
+        assert_that(calling(self.calld_client.relocates.create_from_user).with_args(SOME_CALL_ID, 'wrong-destination'),
                     raises(CalldError).matching(has_properties({
                         'status_code': 400,
                         'error_id': 'invalid-data',
@@ -357,10 +361,10 @@ class TestCreateUserRelocate(TestRelocates):
     def test_given_token_without_user_when_relocate_then_400(self):
         token = 'some-token'
         self.auth.set_token(MockUserToken(token, user_uuid=None))
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
         assert_that(
-            calling(calld.relocates.create_from_user).with_args(
+            calling(self.calld_client.relocates.create_from_user).with_args(
                 SOME_CALL_ID,
                 'line',
                 {'line_id': SOME_LINE_ID}
@@ -373,10 +377,10 @@ class TestCreateUserRelocate(TestRelocates):
     def test_given_no_channel_when_relocate_then_403(self):
         user_uuid = SOME_USER_UUID
         token = self.given_user_token(user_uuid)
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
         assert_that(
-            calling(calld.relocates.create_from_user).with_args(
+            calling(self.calld_client.relocates.create_from_user).with_args(
                 SOME_CALL_ID,
                 'line',
                 {'line_id': SOME_LINE_ID}
@@ -391,10 +395,10 @@ class TestCreateUserRelocate(TestRelocates):
         user_uuid = SOME_USER_UUID
         token = self.given_user_token(user_uuid)
         relocated_channel_id, initiator_channel_id = self.given_bridged_call_stasis()
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
         assert_that(
-            calling(calld.relocates.create_from_user).with_args(
+            calling(self.calld_client.relocates.create_from_user).with_args(
                 initiator_channel_id,
                 'line',
                 {'line_id': SOME_LINE_ID}
@@ -410,10 +414,10 @@ class TestCreateUserRelocate(TestRelocates):
         line_id = SOME_LINE_ID
         token = self.given_user_token(user_uuid)
         relocated_channel_id, initiator_channel_id = self.given_bridged_call_stasis(callee_uuid=user_uuid)
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
         assert_that(
-            calling(calld.relocates.create_from_user).with_args(
+            calling(self.calld_client.relocates.create_from_user).with_args(
                 initiator_channel_id,
                 'line',
                 {'line_id': line_id}
@@ -430,10 +434,10 @@ class TestCreateUserRelocate(TestRelocates):
         token = self.given_user_token(user_uuid)
         relocated_channel_id, initiator_channel_id = self.given_bridged_call_stasis(callee_uuid=user_uuid)
         self.confd.set_users(MockUser(uuid=user_uuid))
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
         assert_that(
-            calling(calld.relocates.create_from_user).with_args(
+            calling(self.calld_client.relocates.create_from_user).with_args(
                 initiator_channel_id,
                 'line',
                 {'line_id': line_id}
@@ -452,10 +456,10 @@ class TestCreateUserRelocate(TestRelocates):
         initiator_channel.setChannelVar(variable='XIVO_USERUUID', value=user_uuid)
         self.confd.set_users(MockUser(uuid=user_uuid, line_ids=[line_id]))
         self.confd.set_lines(MockLine(id=line_id, name='recipient_autoanswer@local', protocol='local', context=SOME_CONTEXT))
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
         assert_that(
-            calling(calld.relocates.create_from_user).with_args(
+            calling(self.calld_client.relocates.create_from_user).with_args(
                 initiator_channel.id,
                 'line',
                 {'line_id': line_id}
@@ -468,10 +472,10 @@ class TestCreateUserRelocate(TestRelocates):
     def test_given_relocate_started_when_relocate_again_then_409(self):
         relocate, user_uuid, destination, location = self.given_ringing_user_relocate()
         token = self.given_user_token(user_uuid)
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
         assert_that(
-            calling(calld.relocates.create_from_user).with_args(
+            calling(self.calld_client.relocates.create_from_user).with_args(
                 relocate['initiator_call'],
                 destination,
                 location,
@@ -489,9 +493,9 @@ class TestCreateUserRelocate(TestRelocates):
         token = self.given_user_token(user_uuid)
         relocated_channel_id, initiator_channel_id = self.given_bridged_call_stasis(callee_uuid=user_uuid)
         events = self.bus.accumulator('calls.relocate.*')
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
-        relocate = calld.relocates.create_from_user(initiator_channel_id, 'line', {'line_id': line_id})
+        relocate = self.calld_client.relocates.create_from_user(initiator_channel_id, 'line', {'line_id': line_id})
 
         until.assert_(
             self.assert_relocate_is_completed,
@@ -532,9 +536,9 @@ class TestCreateUserRelocate(TestRelocates):
         token = self.given_user_token(user_uuid)
         relocated_channel_id, initiator_channel_id = self.given_bridged_call_not_stasis(callee_uuid=user_uuid)
         events = self.bus.accumulator('calls.relocate.*')
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
-        relocate = calld.relocates.create_from_user(initiator_channel_id, 'line', {'line_id': line_id})
+        relocate = self.calld_client.relocates.create_from_user(initiator_channel_id, 'line', {'line_id': line_id})
 
         until.assert_(
             self.assert_relocate_is_completed,
@@ -574,10 +578,10 @@ class TestCreateUserRelocate(TestRelocates):
         self.confd.set_lines(MockLine(id=line_id, context='local'))
         token = self.given_user_token(user_uuid)
         relocated_channel_id, initiator_channel_id = self.given_bridged_call_stasis(callee_uuid=user_uuid)
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
         assert_that(
-            calling(calld.relocates.create_from_user).with_args(
+            calling(self.calld_client.relocates.create_from_user).with_args(
                 initiator_channel_id,
                 'mobile',
             ),
@@ -593,9 +597,9 @@ class TestCreateUserRelocate(TestRelocates):
         self.confd.set_lines(MockLine(id=line_id, context='local'))
         token = self.given_user_token(user_uuid)
         relocated_channel_id, initiator_channel_id = self.given_bridged_call_stasis(callee_uuid=user_uuid)
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
-        relocate = calld.relocates.create_from_user(initiator_channel_id, 'mobile')
+        relocate = self.calld_client.relocates.create_from_user(initiator_channel_id, 'mobile')
 
         until.assert_(
             self.assert_relocate_is_completed,
@@ -645,9 +649,9 @@ class TestCreateUserRelocate(TestRelocates):
     def test_given_relocate_ringing_when_api_cancel_then_relocate_cancelled(self):
         relocate, user_uuid, _, __ = self.given_ringing_user_relocate()
         token = self.given_user_token(user_uuid)
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
-        calld.relocates.cancel_from_user(relocate['uuid'])
+        self.calld_client.relocates.cancel_from_user(relocate['uuid'])
 
         def relocate_cancelled():
             assert_that(relocate['relocated_call'], self.c.is_talking(), 'relocated not talking')
@@ -683,10 +687,10 @@ class TestCreateUserRelocate(TestRelocates):
     def test_given_relocate_waiting_relocated_when_api_cancel_then_400(self):
         relocate, user_uuid = self.given_waiting_relocated_user_relocate()
         token = self.given_user_token(user_uuid)
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
         assert_that(
-            calling(calld.relocates.cancel_from_user).with_args(
+            calling(self.calld_client.relocates.cancel_from_user).with_args(
                 relocate['uuid'],
             ),
             raises(CalldError).matching(has_properties({
@@ -697,9 +701,9 @@ class TestCreateUserRelocate(TestRelocates):
     def test_given_relocate_completion_api_when_api_complete_then_relocate_completed(self):
         relocate, user_uuid = self.given_answered_user_relocate()
         token = self.given_user_token(user_uuid)
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
-        calld.relocates.complete_from_user(relocate['uuid'])
+        self.calld_client.relocates.complete_from_user(relocate['uuid'])
 
         until.assert_(
             self.assert_relocate_is_completed,
@@ -713,9 +717,9 @@ class TestCreateUserRelocate(TestRelocates):
     def test_given_relocate_waiting_completion_when_api_cancel_then_relocate_cancelled(self):
         relocate, user_uuid = self.given_answered_user_relocate()
         token = self.given_user_token(user_uuid)
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
-        calld.relocates.cancel_from_user(relocate['uuid'])
+        self.calld_client.relocates.cancel_from_user(relocate['uuid'])
 
         def relocate_cancelled():
             assert_that(relocate['relocated_call'], self.c.is_talking(), 'relocated not talking')
@@ -763,10 +767,10 @@ class TestCreateUserRelocate(TestRelocates):
     def test_given_ringing_relocate_when_api_complete_then_400(self):
         relocate, user_uuid, _, __ = self.given_ringing_user_relocate()
         token = self.given_user_token(user_uuid)
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
         assert_that(
-            calling(calld.relocates.complete_from_user).with_args(
+            calling(self.calld_client.relocates.complete_from_user).with_args(
                 relocate['uuid'],
             ),
             raises(CalldError).matching(has_properties({
@@ -778,13 +782,13 @@ class TestCreateUserRelocate(TestRelocates):
         mobile_channel, other_channel, mobile_user_uuid = self.given_mobile_call()
         token = self.given_user_token(mobile_user_uuid)
         line_id = SOME_LINE_ID
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
         self.confd.set_users(MockUser(uuid=mobile_user_uuid, line_ids=[line_id]))
         self.confd.set_lines(MockLine(id=line_id, name='recipient_autoanswer@local', protocol='local', context=SOME_CONTEXT))
         destination = 'line'
         location = {'line_id': line_id}
 
-        relocate = calld.relocates.create_from_user(mobile_channel, destination, location)
+        relocate = self.calld_client.relocates.create_from_user(mobile_channel, destination, location)
 
         until.assert_(
             self.assert_relocate_is_completed,
@@ -800,14 +804,14 @@ class TestCreateUserRelocate(TestRelocates):
         initiator_channel, callee_channel = self.given_bridged_call_stasis(caller_uuid=user_uuid)
         token = self.given_user_token(user_uuid)
         line_id = SOME_LINE_ID
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
         self.confd.set_users(MockUser(uuid=user_uuid, line_ids=[line_id], mobile='recipient_autoanswer'))
         self.confd.set_lines(MockLine(id=line_id, name='recipient_autoanswer@local', protocol='local', context='local'))
 
-        relocate = calld.relocates.create_from_user(initiator_channel, destination='mobile')
+        relocate = self.calld_client.relocates.create_from_user(initiator_channel, destination='mobile')
 
         def relocate_finished(relocate):
-            assert_that(calling(calld.relocates.get_from_user).with_args(relocate['uuid']),
+            assert_that(calling(self.calld_client.relocates.get_from_user).with_args(relocate['uuid']),
                         raises(CalldError).matching(has_properties({
                             'status_code': 404,
                             'error_id': 'no-such-relocate',
@@ -823,7 +827,7 @@ class TestCreateUserRelocate(TestRelocates):
         destination = 'line'
         location = {'line_id': line_id}
 
-        relocate = calld.relocates.create_from_user(initiator_channel, destination, location)
+        relocate = self.calld_client.relocates.create_from_user(initiator_channel, destination, location)
 
         until.assert_(
             self.assert_relocate_is_completed,
@@ -841,15 +845,15 @@ class TestCreateUserRelocate(TestRelocates):
         self.confd.set_lines(MockLine(id=line_id, name='ring@local', protocol='local', context=SOME_CONTEXT))
         token = self.given_user_token(user_uuid)
         relocated_channel_id, initiator_channel_id = self.given_bridged_call_stasis(callee_uuid=user_uuid)
-        calld = self.make_calld(token)
+        self.calld_client.set_token(token)
 
-        relocate = calld.relocates.create_from_user(initiator_channel_id, 'line', {'line_id': line_id}, timeout=1)
+        relocate = self.calld_client.relocates.create_from_user(initiator_channel_id, 'line', {'line_id': line_id}, timeout=1)
 
         def relocate_cancelled():
             assert_that(relocate['relocated_call'], self.c.is_talking(), 'relocated not talking')
             assert_that(relocate['initiator_call'], self.c.is_talking(), 'initiator not talking')
             assert_that(relocate['recipient_call'], self.c.is_hungup(), 'recipient not hungup')
-            assert_that(calling(calld.relocates.get_from_user).with_args(relocate['uuid']),
+            assert_that(calling(self.calld_client.relocates.get_from_user).with_args(relocate['uuid']),
                         raises(CalldError).matching(has_properties({
                             'status_code': 404,
                             'error_id': 'no-such-relocate',