+++ wazo_calld/plugin_helpers/ari_.py
@@ -1,6 +1,7 @@
-# Copyright 2015-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2015-2020 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
+import logging
 import time
 import json
 
@@ -10,6 +11,29 @@ from .exceptions import (
     NotEnoughChannels,
     TooManyChannels,
 )
+
+logger = logging.getLogger(__name__)
+
+
+def set_channel_var_sync(channel, var, value, bypass_stasis=False):
+    # TODO remove this when Asterisk gets fixed to set var synchronously
+    def get_value():
+        try:
+            return channel.getChannelVar(variable=var)['value']
+        except ARINotFound as e:
+            if e.original_error.response.reason == 'Variable Not Found':
+                return None
+            raise
+
+    channel.setChannelVar(variable=var, value=value, bypassStasis=bypass_stasis)
+    for _ in range(20):
+        if get_value() == value:
+            return
+
+        logger.debug('waiting for a setvar to complete')
+        time.sleep(0.01)
+
+    raise Exception('failed to set channel variable {}={}'.format(var, value))
 
 
 class GlobalVariableAdapter:
@@ -184,6 +208,13 @@ class Channel:
         try:
             on_hold = self._get_var('XIVO_ON_HOLD')
             return on_hold == '1'
+        except ARINotFound:
+            return False
+
+    def muted(self):
+        try:
+            muted = self._get_var('WAZO_CALL_MUTED')
+            return muted == '1'
         except ARINotFound:
             return False