+++ wazo_calld/plugins/calls/event.py
@@ -1,4 +1,4 @@
-# Copyright 2016-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2016-2020 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import iso8601
@@ -66,3 +66,24 @@ class ConnectCallEvent(StartCallEvent):
         if len(event['args']) < 3:
             raise InvalidConnectCallEvent()
         return event['args'][2]
+
+
+class _BaseEvent:
+
+    required_acl = 'events.{}'
+
+    def marshal(self):
+        return self._body
+
+
+class _BaseCallItemEvent(_BaseEvent):
+
+    def __init__(self, call):
+        self.routing_key = self.routing_key.format(call['call_id'])
+        self.required_acl = self.required_acl.format(self.routing_key)
+        self._body = call
+
+
+class CallUpdated(_BaseCallItemEvent):
+    name = 'call_updated'
+    routing_key = 'calls.{}.updated'