+++ wazo_calld/plugins/calls/http.py
@@ -1,4 +1,4 @@
-# Copyright 2015-2019 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2015-2020 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 import logging
@@ -90,6 +90,54 @@ class CallResource(AuthResource):
         self.calls_service.hangup(call_id)
 
         return None, 204
+
+
+class CallMuteStartResource(AuthResource):
+
+    def __init__(self, calls_service):
+        self.calls_service = calls_service
+
+    @required_acl('calld.calls.{call_id}.mute.start.update')
+    def put(self, call_id):
+        self.calls_service.mute(call_id)
+        return '', 204
+
+
+class CallMuteStopResource(AuthResource):
+
+    def __init__(self, calls_service):
+        self.calls_service = calls_service
+
+    @required_acl('calld.calls.{call_id}.mute.stop.update')
+    def put(self, call_id):
+        self.calls_service.unmute(call_id)
+        return '', 204
+
+
+class MyCallMuteStartResource(AuthResource):
+
+    def __init__(self, auth_client, calls_service):
+        self.auth_client = auth_client
+        self.calls_service = calls_service
+
+    @required_acl('calld.users.me.calls.{call_id}.mute.start.update')
+    def put(self, call_id):
+        user_uuid = get_token_user_uuid_from_request(self.auth_client)
+        self.calls_service.mute_user(call_id, user_uuid)
+        return '', 204
+
+
+class MyCallMuteStopResource(AuthResource):
+
+    def __init__(self, auth_client, calls_service):
+        self.auth_client = auth_client
+        self.calls_service = calls_service
+
+    @required_acl('calld.users.me.calls.{call_id}.mute.stop.update')
+    def put(self, call_id):
+        user_uuid = get_token_user_uuid_from_request(self.auth_client)
+        self.calls_service.unmute_user(call_id, user_uuid)
+        return '', 204
 
 
 class MyCallResource(AuthResource):