+++ wazo_calld/plugins/calls/tests/test_services.py
@@ -1,4 +1,4 @@
-# Copyright 2017-2018 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2017-2020 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from collections import defaultdict
@@ -15,7 +15,7 @@ from ..services import CallsService
 class Testclassname(TestCase):
 
     def setUp(self):
-        self.services = CallsService(Mock(), Mock(), Mock(), Mock(), Mock())
+        self.services = CallsService(Mock(), Mock(), Mock(), Mock(), Mock(), Mock())
 
     def test_given_no_chan_variables_when_make_call_from_ami_event_then_call_has_none_values(self):
         event = defaultdict(str)